requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15
django-cors-headers>=4.4
django-environ>=0.11
celery>=5.4
redis>=5.0
mistralai>=1.0
requests>=2.32
pandas>=2.2
pandas-ta>=0.3.14b
python-dateutil>=2.9
psycopg2-binary>=2.9
gunicorn>=22.0
whitenoise>=6.7
PyPDF2>=3.0
python-docx>=1.0
openpyxl>=3.1
